requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0
opencv-python>=4.5.0
librosa>=0.8.0
pillow>=8.0.0
matplotlib>=3.3.0
seaborn>=0.11.0
sktime>=0.8.0
albumentations>=1.0.0
pyarrow>=5.0.0
pydantic>=1.8.0
python-magic>=0.4.24
moviepy>=1.0.3
soundfile>=0.10.0
joblib>=1.0.0
tqdm>=4.62.0 